requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncmy==0.2.9
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.3
Flask==3.0.3
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
validators==0.34.0
Werkzeug==3.0.6
zipp==3.20.2